add run task to community build Main
